requirements.txt
Django==1.9.5
Django-Select2==4.3.2
Pillow==3.1.1
Unidecode==0.04.19
argparse==1.4.0
cmsplugin-filer==1.0.1
dj-database-url==0.4.0
django-admin-jqueryui==1.8.22
django-appconf==1.0.1
django-autocomplete-light==3.1.5
django-classy-tags==0.7.2
django-cms==3.2.3
django-filer==1.1.1
djangocms-forms==0.1.17
django-formtools==1.0
django-inline-ordering==1.0.2
django-mobile==0.7.0
django-mptt==0.8.3
django-polymorphic==0.8.1
django-recaptcha==1.0.5
django-reversion==1.8.7
django-sekizai==0.9.0
django-simple-math-captcha==1.0.7
django-tagging==0.4.1
django-treebeard==4.0
djangocms-admin-style==1.1.0
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-flash==0.3.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-installer==0.8.7
djangocms-link==1.7.2
djangocms-picture==1.0.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.0
djangocms-video==1.0.0
easy-thumbnails==2.3
html5lib==0.9999999
imagestore==2.9.1
jsonfield==1.0.3
pytz==2016.3
six==1.10.0
sorl-thumbnail==12.3
swapper==0.3.0
tzlocal==1.2.2
wsgiref==0.1.2
mysql-python
psycopg2